+++ Classes/Service/AbstractService.php
@@ -5,8 +5,7 @@ namespace T3v\T3vCore\Service;
 
 use TYPO3\CMS\Core\SingletonInterface;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
-use TYPO3\CMS\Extbase\Configuration\ConfigurationManagerInterface;
-use TYPO3\CMS\Extbase\Object\Exception;
+use TYPO3\CMS\Extbase\Configuration\ConfigurationManager;
 use TYPO3\CMS\Extbase\Object\ObjectManager;
 use TYPO3\CMS\Frontend\ContentObject\ContentObjectRenderer;
 
@@ -21,6 +20,7 @@ abstract class AbstractService implements SingletonInterface
      * Gets the object manager.
      *
      * @return ObjectManager The object manager
+     * @deprecated `ObjectManager` is deprecated since TYPO3 10.4 and will be removed in version 12
      */
     protected static function getObjectManager(): ObjectManager
     {
@@ -30,23 +30,21 @@ abstract class AbstractService implements SingletonInterface
     /**
      * Gets the configuration manager.
      *
-     * @return ConfigurationManagerInterface The configuration manager
-     * @throws Exception
+     * @return ConfigurationManager The configuration manager
      */
-    protected static function getConfigurationManager(): ConfigurationManagerInterface
+    protected static function getConfigurationManager(): ConfigurationManager
     {
-        return self::getObjectManager()->get(ConfigurationManagerInterface::class);
+        return GeneralUtility::makeInstance(ConfigurationManager::class);
     }
 
     /**
      * Gets the content object renderer.
      *
      * @return ContentObjectRenderer The content object renderer
-     * @throws Exception
      */
     protected static function getContentObjectRenderer(): ContentObjectRenderer
     {
-        return self::getObjectManager()->get(ContentObjectRenderer::class);
+        return GeneralUtility::makeInstance(ContentObjectRenderer::class);
     }
 
     /**
@@ -55,7 +53,6 @@ abstract class AbstractService implements SingletonInterface
      * Alias for the `getContentObjectRenderer` function.
      *
      * @return ContentObjectRenderer The content object, respectively the renderer
-     * @throws Exception
      */
     protected static function getContentObject(): ContentObjectRenderer
     {