+++ Classes/Service/LocalizationService.php
@@ -7,6 +7,8 @@ use TYPO3\CMS\Core\Context\Context;
 use TYPO3\CMS\Core\Context\Exception\AspectNotFoundException;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 
+use function is_object;
+
 /**
  * The localization service class.
  *